% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_partialr2.r
\name{calc_partialr2}
\alias{calc_partialr2}
\title{Calculate partial \ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}}}
\usage{
calc_partialr2(output_simple, output_complex)
}
\arguments{
\item{output_simple}{A list outputted from the \code{fit_punc_model} function
(the simpler model)}

\item{output_complex}{A list outputted from the \code{fit_punc_model} function
(the more complex model)}
}
\value{
This function returns the partial
\ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}} value.
}
\description{
This function calculates the partial
\ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}} of a phylogenetic
regression model. In the context of models for detecting punctuated
evolution, this value represents the proportion of the deviation from the
strict molecular (or morphological) clock that is attributable to net
speciation or gene duplication events. Specifically, the partial
\ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}} roughly represents
how much the sum of squared errors (SSE) decreases after including node
count in the model. So, the more complex model has node count as the only
additional parameter (e.g., \code{path ~ node} or \code{path ~ time + node})
compared to the simpler model (e.g., \code{path ~ 1} or \code{path ~ time}).
}
\references{
\href{https://doi.org/10.1080/00031305.1994.10476036}{Anderson-Sprecher (1994)}
}
\author{
Kevin Surya
}
