% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_r2.r
\name{calc_r2}
\alias{calc_r2}
\title{Calculate \ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}}}
\usage{
calc_r2(output_null, output_alt)
}
\arguments{
\item{output_null}{A list outputted from the \code{fit_punc_model} function (the
null model, typically a phylogenetic mean-only model)}

\item{output_alt}{A list outputted from the \code{fit_punc_model} function (the
alternative model)}
}
\value{
This function returns the
\ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}} value.
}
\description{
This function calculates the
\ifelse{html}{\out{<i>R<sup>2</sup></i>}}{\eqn{R^2}} of a phylogenetic
regression model. This value represents how much the predictors in the
model explain (or predict) the variation in the response variable.
}
\references{
\href{https://handle.nal.usda.gov/10113/IND43966364}{Wright (1921)}
}
\author{
Kevin Surya
}
