% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vif.r
\name{calc_vif}
\alias{calc_vif}
\title{Calculate the variance inflation factors (\eqn{VIFs})}
\usage{
calc_vif(output_reg, file)
}
\arguments{
\item{output_reg}{A list outputted from the \code{fit_punc_model} function}

\item{file}{Output file directory}
}
\value{
This function outputs a text file containing the VIF for each
predictor in the model.
}
\description{
This function prints the variance inflation factors
(\eqn{VIFs}) calculated using the \code{vif} function in the \code{R} package \code{car}
to a text file. A \eqn{VIF > 10} indicates significant multicollinearity.
It means that the standard error for the beta or coefficient of the
predictor in question is more than ten times higher than if the predictor
is independent of the other predictors.
}
\references{
\href{https://link-springer-com/book/10.1007/978-1-4614-1353-0}{Vittinghoff et al. (2011)}
\href{https://socialsciences.mcmaster.ca/jfox/Books/Companion/}{Fox and Weisberg (2019)}
}
\author{
Kevin Surya
}
