% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.r
\name{compare_models}
\alias{compare_models}
\title{Compare model fits with \eqn{BIC}}
\usage{
compare_models(models, file)
}
\arguments{
\item{models}{A list of models to compare (e.g.,
\code{models <- list(model1, model2, model3)}) where each model is an output
from the \code{fit_punc_model} function; it is preferable to order the models
in the list from the simplest to the most complex one}

\item{file}{Output file directory}
}
\value{
This function outputs a text file containing a \eqn{BIC} table.
}
\description{
This function compares model fits with the Bayesian Information
Criterion (\eqn{BIC}). The model with the lowest \eqn{BIC} has the highest
maximum likelihood (or roughly the fit of the model to the data),
accounting for model complexity. A \eqn{BIC} difference between two models
(\eqn{\DeltaBIC}) is equal to the Bayes Factor. A \eqn{\DeltaBIC > 2}
represents at least a 100 times difference in the likelihoods.
}
\references{
\href{https://doi.org/10.1214/aos/1176344136}{Schwarz (1978)};
\href{https://doi.org/10.1080/01621459.1995.10476572}{Kass and Raftery (1995)}
}
\author{
Kevin Surya
}
