% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dmat.r
\name{create_dmat}
\alias{create_dmat}
\title{Create the normalizing matrix \eqn{D}}
\usage{
create_dmat(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo}}
}
\value{
This function returns a \eqn{D} matrix.
}
\description{
This function creates the normalizing matrix \eqn{D} by taking
the Cholesky decomposition of the inverse of the phylogenetic
variance-covariance matrix. This process is called Cholesky whitening. As
far as I know, it requires fewer matrix calculations than some methods
that involve singular value decomposition (SVD).
}
\references{
\href{https://doi.org/10.1080/00031305.2016.1277159}{Kessy et al. (2018)}
}
\author{
Kevin Surya
}
