% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_dinosaur}
\alias{data_dinosaur}
\title{Dinosaur phenotypic trait dataset}
\format{
A data frame with 402 rows and 2 variables:
\describe{
\item{\code{log10_mass}}{Log10-transformed body mass estimate (kg);
species average}
\item{\code{clade}}{Ornithischia, Sauropodomorpha, or Theropoda}
}
}
\source{
\href{https://doi.org/10.1371/journal.pbio.1001853}{Benson et al. (2014)};
\href{https://doi.org/10.1111/pala.12329}{Benson et al. (2017)}
}
\usage{
data_dinosaur
}
\description{
This dataset contains the body masses and clades of 402
Mesozoic dinosaur taxa.
}
\note{
Benson et al. (2014) and Benson et al. (2017) predicted dinosaur adult
body masses using estimated relationships between body mass and limb
skeletal measurements across living tetrapods. See their manuscripts for
more details.

The original dataset contains 993 rows. I removed taxa without body mass
estimates and those not present in the timetree.
}
\keyword{datasets}
