% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_mammal}
\alias{data_mammal}
\title{Mammal phenotypic trait dataset}
\format{
A data frame with 880 rows and 2 variables:
\describe{
\item{\code{log10_mass}}{Log10-transformed body mass (kg); species average}
\item{\code{posture}}{Plantigrady (\code{P}: walking on the whole foot), digitigrady
(\code{D}: walking on toes), or unguligrady (\code{U}: walking on hooved tiptoes)}
}
}
\source{
\href{https://doi.org/10.1073/pnas.1814329116}{Kubo et al. (2019)}
}
\usage{
data_mammal
}
\description{
This dataset contains the body masses and foot postures of 880
terrrestrial mammal species.
}
\keyword{datasets}
