% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_vcv.r
\name{decomp_vcv}
\alias{decomp_vcv}
\title{Decompose the variance-covariance matrix}
\usage{
decomp_vcv(tree, lambda = 1, est_lambda = FALSE)
}
\arguments{
\item{tree}{An object of class \code{phylo}}

\item{lambda}{Pagel's \eqn{\lambda}, a measure of phylogenetic signal}

\item{est_lambda}{If \code{TRUE}, the downstream function(s) estimate Pagel's
\eqn{\lambda}}
}
\value{
This function returns a list comprising a correlation matrix (a
\code{corPagel} object) and a variance weight vector.
}
\description{
This function decomposes the phylogenetic variance-covariance
matrix into its correlation (off-diagonal) and variance components
(on-diagonal). The output from \code{decomp_vcv} is necessary for using the
\code{est_node_density} and \code{fit_punc_model} functions.
}
\references{
\href{https://doi.org/10.1038/44766}{Pagel (1999)};
\href{https://doi.org/10.1086/343873}{Freckleton et al. (2002)};
\href{https://doi.org/10.1093/bioinformatics/bty633}{Paradis and Schliep (2019)}
}
\author{
Kevin Surya
}
