% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_branching_rate.r
\name{est_branching_rate}
\alias{est_branching_rate}
\title{Estimate the average branching rate of every lineage}
\usage{
est_branching_rate(data, root_age, timetree)
}
\arguments{
\item{data}{A data frame with node count in the 2nd column, time in the 3rd,
and taxon names as the row names}

\item{root_age}{The age of the tree root; ensure that the time unit is the
same between the root age and the sampling times}

\item{timetree}{An object of the class \code{phylo}}
}
\value{
This function returns a vector of the net number of branching events
per unit of time.
}
\description{
This function estimates the average branching rate of every
lineage (node count per time duration) in the phylogenetic tree. This
metric's goal is to gauge the variability in the net speciation and or
duplication rates.
}
\references{
\href{https://doi.org/10.1086/588076}{Freckleton et al. (2008)}
}
\author{
Kevin Surya
}
