% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_node_density.r
\name{est_node_density}
\alias{est_node_density}
\title{Test for the node-density artifact}
\usage{
est_node_density(data, vcv_parts)
}
\arguments{
\item{data}{A data frame with path length in the 1st column, node count
in the 2nd, and taxon names as the row names; make sure that the row order
matches the tree tip labels (use the function \code{reorder_data})}

\item{vcv_parts}{A list outputted from the \code{decomp_vcv} function}
}
\value{
This function returns a list of the fitted model and node-density
parameter estimates.
}
\description{
This function checks for the node-density artifact, where
branches in parts of a phylogenetic tree with few tips are shorter than
their actual lengths. This artifact can bias the regression model towards
detecting punctuated evolution even if this is not the case. A symptom of
this artifact would be a positive curvilinear relationship between the
node count and phylogenetic path length (\eqn{\delta > 1}).
}
\references{
\href{https://doi.org/10.1126/science.1083202}{Webster et al. (2003)};
\href{https://doi.org/10.1080/10635150600865567}{Venditti et al. (2006)}
}
\author{
Kevin Surya
}
