% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_punc_contrib.r
\name{est_punc_contrib}
\alias{est_punc_contrib}
\title{Estimate the punctuational contribution to the total divergence}
\usage{
est_punc_contrib(tree, output_reg)
}
\arguments{
\item{tree}{An object of class \code{phylo}}

\item{output_reg}{A list outputted from the \code{fit_punc_model} function}
}
\value{
This function returns the punctuational contribution value.
}
\description{
This function estimates how much of the total divergence,
molecular or morphological, is attributable to speciation or gene
duplication events.
}
\references{
\href{https://doi.org/10.1126/science.1129647}{Pagel et al. (2006)}
}
\author{
Kevin Surya
}
