% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.r
\name{export_data}
\alias{export_data}
\title{Export dataset to a text file}
\usage{
export_data(data, file)
}
\arguments{
\item{data}{A data frame}

\item{file}{Output file directory}
}
\value{
This function outputs a tab-delimited text file containing the
dataset.
}
\description{
This function exports a data frame to a tab-delimited text file.
}
\author{
Kevin Surya
}
