% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.r
\name{export_plot_html}
\alias{export_plot_html}
\title{Export plot to an HTML file}
\usage{
export_plot_html(plot, file, plot_3d = FALSE)
}
\arguments{
\item{plot}{An object of the \code{plotly} class}

\item{file}{Output file directory (save as an .html file)}

\item{plot_3d}{Exports a 3D plot if \code{TRUE}}
}
\value{
This function outputs an object of the \code{plotly} class to an HTML
file.
}
\description{
This function exports an interactive plot (2D or 3D) to an HTML
file.
}
\author{
Kevin Surya
}
