% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.r
\name{export_plot_svg}
\alias{export_plot_svg}
\title{Export plot to a SVG file}
\usage{
export_plot_svg(plot, file, width = 4.75, height = 2.94)
}
\arguments{
\item{plot}{An object of the \code{ggplot} class}

\item{file}{Output file directory}

\item{width}{Plot width (inches)}

\item{height}{Plot height (inches)}
}
\value{
This function outputs an object of the \code{ggplot} class to a SVG file.
}
\description{
This function exports a plot to a SVG file. The default plot
dimension (width/height) follows the golden ratio, 1.618.
}
\references{
\href{https://cran.r-project.org/web/packages/Cairo/index.html}{Urbanek and Horner (2022)}
}
\author{
Kevin Surya
}
