% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tree.r
\name{export_tree}
\alias{export_tree}
\title{Export phylogenetic tree to a file}
\usage{
export_tree(tree, file)
}
\arguments{
\item{tree}{An object of class \code{phylo}}

\item{file}{Output file directory}
}
\value{
This function outputs a NEXUS tree file.
}
\description{
This function exports a phylogenetic tree as a NEXUS-formatted
file; it is an alias of the \code{writeNexus} function from the \code{R} package
\code{phytools}.
}
\references{
\href{https://doi.org/10.1111/j.2041-210X.2011.00169.x}{Revell (2012)}
}
\author{
Kevin Surya
}
