% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallpaddy.r
\docType{package}
\name{fallpaddy}
\alias{fallpaddy}
\title{Detecting punctuated evolution in taxa ranging from dinosaurs to viruses}
\description{
This package contains the tools to simulate, detect, and
visualize punctuated evolution in any clades, from co-occurring species
(e.g., present-day mammals) to serially-sampled viruses (e.g., SARS-CoV-2
genomes) and long-extinct taxa (e.g., Mesozoic dinosaurs).
}
\section{Function directory}{

\itemize{
\item Main functions:
\itemize{
\item Extracting data: \code{get_node_count}, \code{get_path_length}, and
\code{get_sampling_time}
\item Preparing dataset: \code{reorder_data}
\item Editing the phylogenetic variance-covariance matrix:
\code{decomp_vcv} and \code{create_dmat}
\item Detecting punctuated evolution: \code{fit_punc_model}
\item Calculating test statistics: \code{calc_r2}, \code{calc_partialr2},
\code{est_branching_rate} and \code{est_punc_contrib}
\item Comparing models: \code{compare_models}
\item Checking biases and diagnostics: \code{calc_vif}, \code{est_node_density},
\code{plot_diagnostics}, and \code{plot_node_density}
\item Visualizing data and results: \code{export_plot_html},
\code{export_plot_pdf}, \code{export_plot_svg}, \code{plot_data_2d},
\code{plot_data_3d}, \code{plot_punc_model}, \code{plot_tree}, and \code{plot_tree_color}
\item Simulating punctuated evolution: \code{sim_punc}
}
\item Auxillary functions:
\itemize{
\item Importing dataset: \code{import_data}, \code{import_data_bt},
\code{import_data_csv}, and \code{import_data_excel}
\item Importing phylogenetic tree: \code{import_tree_nex} and
\code{import_tree_nwk}
\item Exporting dataset: \code{export_data} and \code{export_data_bt}
\item Exporting tree: \code{export_tree_nex} and \code{export_tree_nwk}
\item Printing outputs: \code{print_node_density} and \code{print_punc_model}
}
}
}

\references{
Surya, K., Gardner, J.D. & Organ, C. Detecting punctuated
evolution in SARS-CoV-2 over the first year of the pandemic.
\emph{Frontiers in Virology}.
\href{https://doi.org/10.3389/fviro.2023.1066147}{Link}.
}
\author{
Kevin Surya
}
