% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_punc_model.r
\name{fit_punc_model}
\alias{fit_punc_model}
\title{Detect punctuated evolution}
\usage{
fit_punc_model(data, vcv_parts, D, model = c("p", "pn", "pt", "ptn"))
}
\arguments{
\item{data}{A data frame with path length in the 1st column, node count in
the 2nd, time in the 3rd (optional), and taxon names as the row names;
make sure that the row order matches the tree tip labels (use the function
\code{reorder_data})}

\item{vcv_parts}{A list outputted from the \code{decomp_vcv} function}

\item{D}{A normalizing matrix D outputted from the \code{create_dmat} function}

\item{model}{Model options:
\itemize{
\item "p": path ~ 1
\item "pn": path ~ node
\item "pt": path ~ time
\item "ptn": path ~ time + node
}}
}
\value{
This function returns a list containing the fitted model (an object
of class \code{gls}), estimated variance, phylogenetically-normalized
residuals, and sum of squared errors (SSE).
}
\description{
This function fits the regression model(s) for detecting the
punctuational effect at branching events.
}
\details{
The goal is to estimate the expected difference in net evolution,
morphological or molecular, between two taxa where one of them has
undergone one additional branching event. A taxon's net divergence is
proxied by its phylogenetic path length or root-to-tip distance. And the
cumulative number of branching events along the path is the node count.

This expected difference should account for the strict clock, morphological
or molecular. When all taxa in the tree are co-occurring (e.g., all
present-day mammal species), the expected path length under the strict
clock is the mean (\code{path ~ 1}). Note that the average should be
phylogenetically-normalized. This mean-only model is the simplest.

Punctuated evolution means that more change accumulates during branching
events. As a result, the evolutionary rate is not clock-like. So, we
hypothesize that node count predicts deviation from the clock
(\code{path ~ node}).

However, the strict clock has a different expectation when the taxa in the
tree are not co-occurring. For example, the taxa are sampled serially like
SARS-CoV-2 genomes during the COVID-19 pandemic or sporadically like the
deposition of dinosaur fossils in ancient sediments. In these cases, the
strict clock expects the amount of divergence to scale proportionally with
time (\verb{path ~ sampling time}).

In such cases, the detection of punctuated evolution becomes the regression
of path length on sampling time and node count (\code{path ~ time + node}).

We can fit more complex models where we allow the degree of punctuational
effect to vary with time (e.g., \code{path ~ time + node + time * node}) or
space (e.g., \code{path ~ time + node * continent}). However, it is best to
create your custom function or write a custom script at this point.
}
\references{
\href{https://doi.org/10.1126/science.1083202}{Webster et al. (2003)};
\href{https://doi.org/10.1126/science.1129647}{Pagel et al. (2006)}
}
\author{
Kevin Surya
}
