% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_count.r
\name{get_node_count}
\alias{get_node_count}
\title{Extract node counts}
\usage{
get_node_count(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo}}
}
\value{
This function returns a vector of node counts.
}
\description{
This function extracts the cumulative number of internal nodes
along the root-to-tip paths.
}
\author{
Kevin Surya
}
