% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_length.r
\name{get_path_length}
\alias{get_path_length}
\title{Extract phylogenetic path lengths}
\usage{
get_path_length(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo}}
}
\value{
This function returns a vector of path lengths.
}
\description{
This function extracts the phylogenetic path lengths or
root-to-tip divergences from a molecular tree or a non-ultrametric
timetree.
}
\author{
Kevin Surya
}
