% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sampling_time.r
\name{get_sampling_time}
\alias{get_sampling_time}
\title{Extract sampling times from tip labels}
\usage{
get_sampling_time(tree, block)
}
\arguments{
\item{tree}{An object of class \code{phylo} with tip labels}

\item{block}{The location of the sampling time in the tip label; viral tip
labels typically contain metadata separated by vertical bars "|"; for
example, if the sampling time is in the third "block" (e.g.,
taxon_name|accession_number|sampling_time|sampling_location), then set
\code{block = 3}.}
}
\value{
This function returns a vector of sampling dates in the decimal year
format.
}
\description{
This function extracts the sampling times from tip labels.
}
\details{
Viral tip labels typically include collection or sampling dates in
the ISO 8601 format (YYYY-MM-DD). Whenever the sampling date is only
available in the year (YYYY) or month of the year format (YYYY-MM), this
function fixes sampling time to the middle of the year or the month. This
step introduces a bias but is necessary for conversion to the decimal year
unit. The significance of this bias depends on the timescale of the tree.
For example, within-year date errors are arguably negligible when the
timescale spans decades.
}
\author{
Kevin Surya
}
