% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.r
\name{import_data}
\alias{import_data}
\title{Import dataset}
\usage{
import_data(file)
}
\arguments{
\item{file}{Input file directory}
}
\value{
This function returns an object of class \code{data.frame}.
}
\description{
This function imports a tab-delimited data frame with headers
and row names.
}
\author{
Kevin Surya
}
