% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.r
\name{import_data_bt}
\alias{import_data_bt}
\title{Import dataset (BayesTraits style)}
\usage{
import_data_bt(file)
}
\arguments{
\item{file}{Input file directory}
}
\value{
This function returns an object of class \code{data.frame}.
}
\description{
This function imports a tab-delimited data frame without
headers (BayesTraits style).
}
\author{
Kevin Surya
}
