% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.r
\name{import_data_csv}
\alias{import_data_csv}
\title{Import a .csv file}
\usage{
import_data_csv(file)
}
\arguments{
\item{file}{Input file directory}
}
\value{
This function returns an object of class \code{data.frame}.
}
\description{
This function is an alias of \code{read.csv}.
}
\author{
Kevin Surya
}
