% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.r
\name{import_data_excel}
\alias{import_data_excel}
\title{Import a .xls or .xlsx file}
\usage{
import_data_excel(file, sheet)
}
\arguments{
\item{file}{Input file directory}

\item{sheet}{Excel sheet}
}
\value{
This function returns an object of class \code{tbl_df}.
}
\description{
This function is an alias of \code{read_excel} from the \code{R} package
\code{readxl}.
}
\author{
Kevin Surya
}
