% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tree.r
\name{import_tree_nwk}
\alias{import_tree_nwk}
\title{Import phylogenetic tree (Newick format)}
\usage{
import_tree_nwk(file)
}
\arguments{
\item{file}{Input file directory}
}
\value{
This function returns an object of class \code{phylo}.
}
\description{
This function imports a phylogenetic tree file written in the
Newick format; it is an alias of the \code{read.newick} function from the \code{R}
package \code{phytools}.
}
\references{
\href{https://doi.org/10.1111/j.2041-210X.2011.00169.x}{Revell (2012)}
}
\author{
Kevin Surya
}
