% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.r
\name{plot_data_3d}
\alias{plot_data_3d}
\title{Create a 3D interactive plot}
\usage{
plot_data_3d(data, group = FALSE, size = 1.25)
}
\arguments{
\item{data}{A data frame with taxon name in the row name, path length in the
1st column, node count in the 2nd, time in the 3rd, and group assignment
in the 4th (optional)}

\item{group}{If \code{TRUE}, the colors of the data points represent the group
assignment instead of the node count}

\item{size}{Size of data point}
}
\value{
This function returns an object of the class \code{plotly} class.
}
\description{
This function creates a 3D interactive plot of the data.
}
\author{
Kevin Surya
}
