% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.r
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plot regression diagnostics}
\usage{
plot_diagnostics(output_reg)
}
\arguments{
\item{output_reg}{A list outputted from the \code{fit_punc_model} function}
}
\value{
This function returns an object of the \code{ggplot} class.
}
\description{
This function plots the distribution of the
phylogenetically-normalized residuals to check if it meets the assumption
of normality. Another panel in the plot is the residuals vs. fitted
values. It is typically helpful for diagnosing model adequacy and the
equal variance assumption. There should be little to no trend and no
"funneling" pattern in the residuals vs. fitted plot.
}
\references{
\href{https://doi.org/10.1111/j.1420-9101.2009.01757.x}{Freckleton (2009)};
\href{https://doi.org/10.1007/978-3-662-43550-2_6}{Mundry (2014)}
}
\author{
Kevin Surya
}
