% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_node_density.r
\name{plot_node_density}
\alias{plot_node_density}
\title{Plot the node-density test}
\usage{
plot_node_density(data, output_nd, unit)
}
\arguments{
\item{data}{A data frame with path length in the 1st column and node count
in the 2nd}

\item{output_nd}{A list outputted from the \code{est_node_density} function}

\item{unit}{Branch length unit (e.g., subs/site and mya)}
}
\value{
This function returns an object of the \code{ggplot} class.
}
\description{
This function plots the node-density test, testing for a
curvilinear relationship between the node count and path length.
}
\references{
\href{https://doi.org/10.1126/science.1083202}{Webster et al. (2003)};
\href{https://doi.org/10.1080/10635150600865567}{Venditti et al. (2006)}
}
\author{
Kevin Surya
}
