% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_node_density.r
\name{print_node_density}
\alias{print_node_density}
\title{Print the node-density test output}
\usage{
print_node_density(output_nd, file)
}
\arguments{
\item{output_nd}{A list outputted from the \code{est_node_density} function}

\item{file}{Output file directory}
}
\value{
This function outputs a text file containing the summary of the
node-density test.
}
\description{
This function saves the summary of the node-density test to a
text file.
}
\author{
Kevin Surya
}
