% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_punc_model.r
\name{print_punc_model}
\alias{print_punc_model}
\title{Print the output of a regression model}
\usage{
print_punc_model(output_reg, file)
}
\arguments{
\item{output_reg}{A list outputted from the \code{fit_punc_model} function}

\item{file}{Output file directory}
}
\value{
This function outputs a text file containing the summary of the
regression model.
}
\description{
This function saves the summary of a regression model to a text
file.
}
\author{
Kevin Surya
}
