% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_data.r
\name{reorder_data}
\alias{reorder_data}
\title{Re-order the dataset to follow the tree}
\usage{
reorder_data(data, tree)
}
\arguments{
\item{data}{A data frame}

\item{tree}{An object of the class \code{phylo}}
}
\value{
This function returns a re-ordered data frame.
}
\description{
This function reorders rows in a data frame to follow the order
of tree tip labels
}
\author{
Kevin Surya
}
