% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_punc.r
\name{sim_punc}
\alias{sim_punc}
\title{Simulate a phylogenetic tree with a specified signature of punctuated
evolution}
\usage{
sim_punc(n, age, rate, punc, sampling = c("extant", "serial_fossil"))
}
\arguments{
\item{n}{Number of tips in the tree}

\item{age}{The total depth of the timetree; the unit is up to the user
(e.g., \code{age = 100} could mean that the age of the most recent common
ancestor is 100 years or 100 million years ago)}

\item{rate}{The maximum net evolutionary rate; the unit is up to the user
(e.g., \code{rate = 0.001} could mean 0.001 DNA substitutions per site per year
or 0.001 kilograms per million year)}

\item{punc}{The proportion of total tree length attributable to
punctuational effects (ranges from 0 [completely gradual] to 1
[completely punctuational])}

\item{sampling}{List of options:
\itemize{
\item "extant": all taxa in the dataset are sampled at the "same" time
(e.g., present-day mammal species)
\item "serial_fossil": taxa are serially-sampled (e.g., SARS-CoV-2
genomes and extinct dinosaur species)
}}
}
\value{
This function returns a list comprising a pair of simulated trees: a
timetree and its corresponding molecular/morphological tree.
}
\description{
This function first simulates a birth-death timetree
(extant-only, serially-sampled, or extinct only) and then the
corresponding molecular or morphological tree that bears the
user-specified signature of punctuated evolution (i.e., the speciational
contribution to the total divergence).
}
\details{
In this simulation, as we turn the knob, increasing the
speciational contribution to the total divergence, the molecular or
morphological evolution becomes less clock-like, and sampling time
predicts divergence less and less. This simulation, however, cannot create
deviations from the strict clock unrelated to punctuation (e.g.,
single-branch or clade-wide rate shifts).
}
\references{
\href{https://doi.org/10.1214/aoms/1177730285}{Kendall (1948)};
\href{https://doi.org/10.1038/44766}{Pagel (1999)};
\href{https://doi.org/10.1126/science.1129647}{Pagel et al. (2006)};
\href{https://doi.org/10.1093/sysbio/syr029}{Stadler (2011)};
\href{https://doi.org/10.1016/j.jtbi.2015.09.005}{Paradis (2015)}
}
\author{
Kevin Surya
}
