% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tree_dinosaur_time}
\alias{tree_dinosaur_time}
\title{Dinosaur timetree}
\format{
An object of class \code{phylo} of length 4
}
\source{
\href{https://doi.org/10.1371/journal.pbio.1001853}{Benson et al. (2014)};
\href{https://doi.org/10.1111/pala.12329}{Benson et al. (2017)};
\href{https://doi.org/10.1038/s41559-017-0454-6}{O'Donovan et al. (2018)}
}
\usage{
tree_dinosaur_time
}
\description{
This time-calibrated phylogenetic tree comprises 402 Mesozoic
dinosaur taxa.
}
\details{
\itemize{
\item \strong{Clade:} Dinosauria (Mesozoic dinosaurs)
\item \strong{Number of tips:} 402 dinosaur taxa
\item \strong{Data:} Cladograms and first and last appearance date estimates
\item \strong{Branch length unit:} Million years
\item \strong{Inference method:} Minimum branch length (mbl) (\code{paleotree})
\item \strong{Rooted?} Yes
\item \strong{Binary or fully-bifurcating?} Yes
}
}
\note{
Regarding the dinosaur timetree, Benson et al. (2014) constructed a
composite cladogram. O'Donovan et al. (2018) acquired this cladogram,
first and last appearance date estimates, and scaled the branch lengths
using the 'mbl' method, enforcing a minimum branch length of one million
years. The tree file is unavailable online, but is available upon request
from the authors.

The timetree from O'Donovan et al. (2018) contains 624 tips. I removed taxa
not present in the body mass dataset. Also, I arbitrarily resolved each
polytomy (or multifurcation) into a series of bifurcations with zero
branch lengths. The resulting bifurcations follow the order in which the
tips appear in the tree.
}
\keyword{datasets}
