% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tree_lepidosaur_mol}
\alias{tree_lepidosaur_mol}
\title{Lepidosaur molecular tree}
\format{
An object of class \code{phylo} of length 4
}
\source{
\href{https://doi.org/10.1186/1471-2148-13-93}{Pyron et al. (2013)}
}
\usage{
tree_lepidosaur_mol
}
\description{
This molecular phylogenetic tree comprises 4,162 lepidosaur
species.
}
\details{
\itemize{
\item \strong{Clade:} Lepidosauria (mostly snakes and lizards)
\item \strong{Number of tips:} 4,162 (4,161 squamate and 1 tuatara species)
\item \strong{Data:} 12 genes (5 mtDNA and 7 nuclear); 12,896 bp
\item \strong{Branch length unit:} Substitutions per site
\item \strong{Inference method:} Maximum likelihood (\code{RAxML})
\item \strong{Rooted?} Yes
\item \strong{Binary or fully-bifurcating?} Yes
}
}
\note{
Pyron et al. (2013) focus on broad taxonomic sampling, resulting in
a tree with a more extensive record of net speciation events than many
other Lepidosauria trees. They found that the terminal branch lengths do
not correlate with the completeness of the molecular data.
}
\keyword{datasets}
