% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tree_zika_time}
\alias{tree_zika_time}
\title{Zika virus timetree}
\format{
An object of class \code{phylo} of length 4
}
\source{
\href{https://doi.org/10.1038/nature22400}{Grubaugh et al. (2017)}
}
\usage{
tree_zika_time
}
\description{
This time-calibrated phylogenetic tree comprises 104 Zika virus
genomes.
}
\details{
\itemize{
\item \strong{Clade:} Zika virus (ZIKV)
\item \strong{Number of tips:} 104 genomes (2013-2016; the Pacific, Brazil,
other South and Central Americas, the Caribbean, and the United States)
\item \strong{Data:} Complete genomes; 10,269 bases
\item \strong{Branch length unit:} Years
\item \strong{Inference method:} Relaxed molecular clock + Bayesian skyline
(\code{BEAST})
\item \strong{Rooted?} Yes
\item \strong{Binary or fully-bifurcating?} Yes
}
}
\note{
The timetree is a maximum clade credibility tree.
}
\keyword{datasets}
