% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.r
\name{plot_data_2d}
\alias{plot_data_2d}
\title{Create a 2D interactive plot}
\usage{
plot_data_2d(data, group = FALSE, size = 1.5, alpha = 1)
}
\arguments{
\item{data}{A data frame with taxon name in the row name, path length in the
1st column, node count in the 2nd, and group assignment in the 3rd
(optional)}

\item{group}{If \code{TRUE}, the colors of the data points represent the group
assignment}

\item{size}{Size of data point}

\item{alpha}{Transparency of data point (0: transparent; 1: opaque)}
}
\value{
This function returns an object of the class \code{plotly}.
}
\description{
This function creates an interactive plot of the data.
}
\author{
Kevin Surya
}
