% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_punc_model.r
\name{plot_punc_model}
\alias{plot_punc_model}
\title{Plot data with or without a fit line}
\usage{
plot_punc_model(
  data,
  output_reg,
  size = 1,
  alpha = 1,
  fit = TRUE,
  type = c("extant_molecular", "extant_morphological", "serial", "fossil")
)
}
\arguments{
\item{data}{A data frame with path length in the 1st column, node count in
the 2nd, time in the 3rd (optional), and group assignment in the 4th
(optional)}

\item{output_reg}{A list outputted from the \code{fit_punc_model} function}

\item{size}{Size of data point}

\item{alpha}{Transparency of data point (0: transparent; 1: opaque)}

\item{fit}{If \code{TRUE}, the plot shows phylogenetically-normalized fitted
lines for \code{path ~ 1}, \code{path ~ node}, and \code{path ~ time} models; for more
complex models, such as those that estimate a fit plane, it is easier to
write a custom script}

\item{type}{List of options:
\itemize{
\item "extant_molecular": all taxa in the dataset are sampled at the
"same" time (e.g., present-day mammal species), and the phylogenetic
tree branch lengths reflect molecular evolution
\item "extant_morphological": all taxa in the dataset are sampled at the
"same" time (e.g., present-day mammal species), and the phylogenetic
tree branch lengths reflect morphological evolution
\item "serial": taxa are serially-sampled, and the timescale is in days
to years (e.g., SARS-CoV-2 genomes)
\item "fossil": taxa are serially-sampled, and the timescale is in
hundreds to millions of years (e.g., extinct dinosaur species)
}}
}
\value{
This function returns an object of the \code{ggplot} class.
}
\description{
This function plots data with or without a fit line derived
from the \code{fit_punc_model} function.
}
\references{
\href{https://doi.org/10.1007/BF01025868}{Freedman and Diaconis (1981)};
\href{https://doi.org/10.1126/science.1083202}{Webster et al. (2003)};
\href{https://doi.org/10.1126/science.1129647}{Pagel et al. (2006)};
\href{https://doi.org/10.1093/ve/vew007}{Rambaut et al. (2016)}
}
\author{
Kevin Surya
}
