% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.r
\name{plot_tree_color}
\alias{plot_tree_color}
\title{Plot phylogenetic tree with colored branches}
\usage{
plot_tree_color(tree, data, unit, rev = FALSE, size = 0.3)
}
\arguments{
\item{tree}{An object of class \code{phylo}}

\item{data}{A data frame with taxon name in the row name and the group
assignment in the 1st column}

\item{unit}{Branch length unit (e.g., subs/site and mya)}

\item{rev}{A logical indicating whether to reverse the x-axis time scale
(defaults to FALSE, e.g., 0, 1, 2, ...)}

\item{size}{Branch thickness}
}
\value{
This function returns an object of the \code{ggtree} class.
}
\description{
This function plots a phylogenetic tree with colors
representing groups
}
\author{
Kevin Surya
}
